Review and cleanup of the code
Move classfileWriter to PostProcessor
Remove UnitInfoLookup, Separate concerns in UnitResult
Remove ExecutorServiceInfo
Set entry sizes in uncompressed jars
Refer to stats from frontendAccess, warn on stats with multi-threaded
   writer
Separate BufferedReporting

Remove unused classes
Add some documentation
